requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
darkdetect==0.8.0
Flask==3.1.2
fonttools==4.61.1
gunicorn==24.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
numpy==2.4.1
packaging==25.0
pillow==12.1.0
pyparsing==3.3.1
pythainlp==5.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.5
python-dotenv==1.0.1
scikit-learn==1.6.1
